Move HalfGreyBackground to CaseStudyBlocks
